Implement the fleet of S3 workers
The S3 workers are responsible for getting tasks from the queue and making
the corresponding S3 requests.  Currently, two types of tasks are implemented.

 * S3TaskTypeWriteFile – put the whole file from local filesystem to S3.
 * S3TaskTypeWriteFilePart – put the 1MB part of the file from local filesystem
                             to S3.

The S3TaskTypeWriteFile tasks are to be used for PostgreSQL default engine,
while S3TaskTypeWriteFilePart tasks are to be used for OrioleDB.

Given that S3 have high latency, we can launch many workers
(orioledb.s3_num_workers) in parallel to acheive the high throughput.

If worker fails to complete the request, the position of in-progress task is
still in the shared memory.  In this case S3 request will be re-tried after
the worker restart.