Implement S3 WAL achiever
We need to archive the WAL files as long as the data files.  This commit
makes OrioleDB an "archive_library", which put the WAL files to S3.
An archive library have to return the result synchronously, and it already
works in the  dedicated archiving process.  Thus, there is no point to assign
S3 workers to archive WAL.  Instead we do the S3 requests to archive WAL
right-away.